==> 01 - Starbucks Capstone Project - Data Preparation.ipynb <==
<h1 style="color: blue;">Starbucks Project Overview</h1>

<h2 style="color: blue;">Starbucks Capstone Challenge</h2>

### Instructions for the project can be found in the Starbucks Project Workspace.

#### Dataset overview

- The program used to create the data simulates how people make purchasing decisions and how those decisions are influenced by promotional offers.
- Each person in the simulation has some hidden traits that influence their purchasing patterns and are associated with their observable traits. People produce various events, including receiving offers, opening offers, and making purchases.
- As a simplification, there are no explicit products to track. Only the amounts of each transaction or offer are recorded.
- There are three types of offers that can be sent: buy-one-get-one (BOGO), discount, and informational. In a BOGO offer, a user needs to spend a certain amount to get a reward equal to that threshold amount. In a discount, a user gains a reward equal to a fraction of the amount spent. In an informational offer, there is no reward, but neither is there a requisite amount that the user is expected to spend. Offers can be delivered via multiple channels.
- The basic task is to use the data to identify which groups of people are most responsive to each type of offer, and how best to present each type of offer.

#### Data Dictionary
##### profile.json
Rewards program users (17000 users x 5 fields)
- gender: (categorical) M, F, O, or null
- age: (numeric) missing value encoded as 118
- id: (string/hash)
- became_member_on: (date) format YYYYMMDD
- income: (numeric)

##### portfolio.json
Offers sent during 30-day test period (10 offers x 6 fields)

- reward: (numeric) money awarded for the amount spent
- channels: (list) web, email, mobile, social
- difficulty: (numeric) money required to be spent to receive reward
- duration: (numeric) time for offer to be open, in days
- offer_type: (string) bogo, discount, informational
- id: (string/hash)

##### transcript.json
Event log (306648 events x 4 fields)

- person: (string/hash)
- event: (string) offer received, offer viewed, transaction, offer completed
- value: (dictionary) different values depending on event type
- offer id: (string/hash) not associated with any "transaction"
- amount: (numeric) money spent in "transaction"
- reward: (numeric) money gained from "offer completed"
- time: (numeric) hours after start of test

#### Import libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from datetime import datetime
import seaborn as sns
import calendar

#### Load the datasets

portfolio = pd.read_json('data/portfolio.json', lines=True)
profile = pd.read_json('data/profile.json', lines=True)
transcript = pd.read_json('data/transcript.json', lines=True)

#### Get the unique channels

unique_channels = list(set(portfolio.channels.explode()))
unique_channels

#### Clean and structure the portfolio dataset

#  for the unique channels separate them and get an structured dataset
portfolio[unique_channels] = list(map(lambda x:  np.in1d(unique_channels, x), portfolio.channels))
portfolio.drop(columns='channels', inplace=True)
portfolio # look at the portfolio

#### Clean and Structure the transcript file

transcript_value = pd.json_normalize(transcript.value)
transcript[transcript_value.columns] = transcript_value
transcript.drop(columns='value', inplace=True)
transcript

profile

#### Merge the transcript file 

df = transcript.merge(profile, left_on='person', right_on='id', how='left')
df = df.merge(portfolio, left_on='offer id', right_on='id', how='left')
df = df.drop(columns=['reward_x', 'id_y', 'id_x']).rename(columns={'reward_y':'reward'})

df = df.sort_values(by=['person', 'time'])
df.drop(columns='offer_id', inplace=True)

%%time
df = pd.concat([df[df.person==p].ffill() for p in df.person.unique()])
df = df.sort_values(['person', 'time', 'offer id'])

#### Make a decomposition of the final dataset and check some basic events

decomposition = df.became_member_on.apply(lambda x: [int(str(x)[:4]), int(str(x)[4:6]), int(str(x)[6:])])
year = decomposition.apply(lambda x: x[0])
month = decomposition.apply(lambda x: x[1])
#df['sin_month'] = df.month.apply(lambda x: np.sin(2*np.pi*x/12))
day = decomposition.apply(lambda x: x[2])
#df['sin_day'] = decomposition.apply(lambda x: np.sin(2*np.pi*x[2]/calendar.monthrange(x[0], x[1])[1]))
min_date = datetime(year.min(), month.min(), day.min())
df['days_since_membership'] = decomposition.apply(lambda x:(datetime(x[0], x[1], x[2]) - min_date).days)
df.drop(columns='became_member_on', inplace=True)
df['offered_channels_count'] = df[['email', 'web', 'mobile', 'social']].sum(axis=1)
df.set_index('offer id', inplace=True)
print('Before Dropping duplicates', df.shape) # basic dataset shape
df = df.drop_duplicates()
print('After Dropping Duplicates', df.shape) # after dropping duplicates

df.head(2)

df.to_csv('dataset.csv', index=True)
