==> 03 - Starbucks Capstone Project - Data Splitting.ipynb <==
!pip install --quiet imbalanced-learn
!pip install --quiet xgboost

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

df = pd.read_csv('dataset.csv', index_col=0)

df.describe(percentiles=[0.05, 0.1, 0.15, 0.25, 0.75, 0.8, 0.9, 0.95])

df = df[(df.age < 90) & (df.income > 25_000) & (df.income < 100_000)]

df.isna().sum()

df.head(2)

person_uniques = df.person.unique()
person_nuniques = len(person_uniques)
person_map = dict(zip(person_uniques, range(1, person_nuniques+1)))

with open('person_dict.json', 'w') as p:
    json.dump(person_map, p)

df.person = df.person.map(person_map)

from imblearn.pipeline import Pipeline
from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
le.fit_transform(df.event)

event_to_labels = dict(zip(le.classes_, range(len(le.classes_))))
labels_to_event = dict(zip(event_to_labels.values(), event_to_labels.keys()))

with open('event_to_labels.json', 'w') as p:
    json.dump(event_to_labels, p)
with open('labels_to_event.json', 'w') as p:
    json.dump(labels_to_event, p)

df[['email', 'social', 'mobile', 'web']] = df[['email', 'social', 'mobile', 'web']].fillna(False).astype(int)

target = 'event'
y = df.groupby('person')[target].shift(-1)
X = df.drop(columns=target)

X = X[~y.isna()]
y.dropna(inplace=True)

tmp = pd.concat([X, y], axis=1)

train, test = train_test_split(tmp, test_size=0.2, random_state=42, stratify=tmp.time)

train.to_csv('train.csv', index=True)
test.to_csv('test.csv', index=True)
