==> 04 - Starbucks Capstone Project - Base Model Test.ipynb <==
!pip install --quiet boto3
!pip install --quiet sagemaker
!pip install --quiet imbalanced-learn
!pip install --quiet xgboost

import pandas as pd
import json

train = pd.read_csv('test.csv', index_col=0)
test  = pd.read_csv('test.csv', index_col=0)

train.head(2)

with open('event_to_labels.json', 'r') as p:
    event_to_labels = json.load(p)

event_to_labels

target = 'event'

X_train = train.drop(columns=target)
y_train = train.get(target).map(event_to_labels)

X_test = train.drop(columns=target)
y_test = train.get(target).map(event_to_labels)

import numpy as np
import pandas as pd
from sklearn.compose import ColumnTransformer
from sklearn.datasets import fetch_openml
from sklearn.feature_selection import SelectPercentile, chi2
from sklearn.impute import SimpleImputer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import RandomizedSearchCV, train_test_split
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import make_column_selector
from imblearn.over_sampling import RandomOverSampler
from imblearn.under_sampling import RandomUnderSampler
from xgboost import XGBClassifier
from imblearn.ensemble import BalancedBaggingClassifier
from imblearn.pipeline import Pipeline
from sklearn.metrics import recall_score, precision_score, classification_report

np.random.seed(42)

numeric_features = make_column_selector(dtype_include=np.number)
numeric_transformer = Pipeline(
    steps=[
        ("imputer", SimpleImputer(strategy="median", add_indicator=True)), 
        ("scaler", StandardScaler())
    ]
)

categorical_features = make_column_selector(dtype_exclude=np.number)
categorical_transformer = Pipeline(
    steps=[
        ("encoder", OneHotEncoder(handle_unknown="ignore", sparse_output=False)),
    ]
)

preprocessor = ColumnTransformer(
    transformers=[
        ("num", numeric_transformer, numeric_features),
        ("cat", categorical_transformer, categorical_features),
    ]
)



ros = RandomOverSampler(random_state=42, sampling_strategy='auto')
rus = RandomUnderSampler(random_state=42, sampling_strategy='auto')
bbc = XGBClassifier(random_state=42)

model = Pipeline(
    steps=[
        ("preprocessor", preprocessor), 
        ("oversample", ros),
        ("undersample", rus),        
        ("model", bbc),
    ]
)

model.fit(X_train, y_train)

y_pred = np.argmax(model.predict_proba(X_train) > 0.47, axis=1)

pd.DataFrame(classification_report(y_train, y_pred, labels=list(event_to_labels.values()), target_names=list(event_to_labels.keys()), output_dict=True)).T

y_pred = np.argmax(model.predict_proba(X_test) > 0.47, axis=1)

pd.DataFrame(classification_report(y_test, y_pred, labels=list(event_to_labels.values()), target_names=list(event_to_labels.keys()), output_dict=True)).T

#### Emulation of the predictor

def model_fn(model):
    # fake model_fn
    return model

def input_fn(request_body, request_content_type):
    print('Start input function')
    if request_content_type=='text/csv':
        data = pd.read_csv(io.StringIO(request_body))
        return data
    if request_content_type=='application/json':
        data = pd.read_json(json.loads(request_body), orient='records')
        return data
    if request_content_type=='application/x-parquet':
        data = pd.read_parquet(io.BytesIO(request_body))
        return data
    raise ValueError('Accept header must be application/json or text/csv or application/x-parquet')

        
def predict_fn(input_object, model):
    print('Start predict function')
    if len(input_object.shape) <= 1:
        input_object = input_object.to_frame().T
    columns = ['offer completed', 'offer received', 'offer viewed', 'transaction']
    predictions = pd.DataFrame(model.predict_proba(input_object), index=input_object.index, columns=columns)
    print('End predict function')
    return predictions

def output_fn(predictions, content_type):
    print('Start output function')
    if content_type=='application/json' or content_type=='application/x-parquet' or content_type=='text/csv':
        result = json.dumps(predictions.to_dict())
        return result
    print('End Output function')

import joblib
import json

model = model_fn(model)

samples = X_train.sample(1).to_json(orient='records')

#samples = X_train.sample(1).to_json(orient='records', lines=True)
data = input_fn(request_body=json.dumps(samples), request_content_type='application/json')
payload = json.dumps(samples)

predictions = predict_fn(data, model)

pd.read_json(output_fn(predictions, content_type='application/json'))
