==> 05 - Starbucks Project - Training - Hyperparameter Tuning Job.ipynb <==
#### Training a XGBoost Classifier

Our approach was to develop a XGBClassifier to predict the "NEXT ACTION"

import boto3
import sagemaker
from sagemaker import get_execution_role
from sagemaker.sklearn import SKLearn
from sagemaker.estimator import Estimator
from sagemaker.inputs import TrainingInput
from sagemaker.tuner import HyperparameterTuner, IntegerParameter, CategoricalParameter, ContinuousParameter
from datetime import datetime
import pytz
import ast

from sagemaker.local import LocalSession

role = get_execution_role()
region = boto3.Session().region_name
sess = sagemaker.Session()

# Local Session
#sess = LocalSession()
#sess.config = {'local': {'local_code':True}}

Bucket = 'sagemaker-us-east-1-254050731868'
Prefix = 'starbucks-capstone'

train_path = f's3://{Bucket}/{Prefix}/dataset/train.csv'
test_path = f's3://{Bucket}/{Prefix}/dataset/test.csv'

#### Preparing the estimator for the hyperparameter tuner job

estimator = Estimator(entry_point='code/train.py',
                      image_uri='683313688378.dkr.ecr.us-east-1.amazonaws.com/sagemaker-scikit-learn:1.2-1-cpu-py3',
                      instance_count=1,
                      instance_type='ml.m5.large',
                      #instance_type='local',
                      framework_version='1.2-1',
                      code_location=f's3://{Bucket}/{Prefix}/model',
                      ouptut_path=f's3://{Bucket}/{Prefix}/model',
                      dependencies=['code/requirements.txt'],
                      role=role,
                      max_run=3600,
                     )

params = {
    "max-depth":6,         # 0-inf (int)
    "min-child-weight": 1, # 0-inf (int)
    "subsample":1,         # 0-1 (float)
    "colsample-bytree":1,  # 0-1 (float)
    "eta": 0.3,            # 0-1 (float)
}

estimator.set_hyperparameters(**params)

#### Launch the hyperparameter tuner job

now = datetime.now()
now_suffix = now.strftime('%b%d%y-%H%M%S').lower()

tuner_args = dict(
    objective_metric_name='test_precision',
    objective_type='Maximize',
    metric_definitions=[{'Name':'test_precision', 'Regex': 'TEST  PRECISION: (-?[0-9\.]+)'}],
    max_jobs=12,
    max_parallel_jobs=3,
    early_stopping_type='Auto'
)

tuner = HyperparameterTuner(
    estimator=estimator,
    hyperparameter_ranges={
        "max-depth"        : IntegerParameter(2, 400),          # 0-inf (int)
        "min-child-weight" : IntegerParameter(1, 400),          # 0-inf (int)
        "subsample"        : ContinuousParameter(0.0001, 0.99), # 0-1 (float)
        "colsample-bytree" : ContinuousParameter(0.0001, 0.99), # 0-1 (float)
        "eta"              : ContinuousParameter(0.0001, 0.99)  # 0-1 (float)
    },
    strategy='Bayesian',
    max_runtime_in_seconds=4*3600,
    base_tuning_job_name='starbucks',
    **tuner_args
)

tuner.fit(
    inputs={'train': train_path, 'test': test_path},
    wait=True,
    job_name=f'tn-stb-event-{now_suffix}'
)

#### Optionally you can train the final model

We prefer this last part to be tested on local to prevent failures of the hyperparameter tunning job rather than training a final model.

We could get the final model from our best hyperparameter tunning job.

now = datetime.now()
now_suffix = now.strftime('%b%d%y-%H%M%S').lower()

estimator.fit(
    wait = True,
    job_name=f'starbucks-{now_suffix}',
    inputs={"train":train_path, "test":test_path}
)
