==> 06 - Starbucks Project - Deploy and Inference.ipynb <==
#### Training a XGBoost Classifier

Our approach was to develop a XGBClassifier to predict the "NEXT ACTION"

import boto3
import sagemaker
import json
from sagemaker import get_execution_role
from sagemaker.sklearn import SKLearnModel
#from sagemaker.estimator import Estimator
from sagemaker import image_uris
from datetime import datetime
import pytz
import ast
from sagemaker.local import LocalSession
from sagemaker.model import Model
from sagemaker.tuner import HyperparameterTuner
from sklearn.metrics import classification_report
import numpy as np

role = get_execution_role()
region = boto3.Session().region_name
sess = sagemaker.Session()
sagemaker_runtime = boto3.client('sagemaker-runtime', region_name=region)

# Local Session
#sess = LocalSession()
#sess.config = {'local': {'local_code':True}}

Bucket = 'sagemaker-us-east-1-254050731868'
Prefix = 'starbucks-capstone'

#### Read The best Model

image_uri = image_uris.retrieve(framework="sklearn", region=region, version="1.2-1")
tuning_job_name = 'tn-stb-event-sep2124-194742'
tuning_job = HyperparameterTuner.attach(tuning_job_name, sagemaker_session=sess)
best_job = tuning_job.best_training_job()
best_hyperparameters = tuning_job.best_estimator().hyperparameters()
model_data = tuning_job.best_estimator().model_data
print('model data: ', model_data)
print('Best Job: ', best_job)
print('Best HyperParameters' ,best_hyperparameters)

#### Deploying the model

model = SKLearnModel(
    model_data=model_data,
    framework_version='1.2-1',
    role=role,
    sagemaker_session=sess,
    dependencies=['code/requirements.txt'],
    entry_point='code/train.py'
)

predictor = model.deploy(initial_instance_count=1, instance_type='ml.t2.medium', endpoint_name='ep-starbucks')

import pandas as pd

train_path = f's3://{Bucket}/{Prefix}/dataset/train.csv'
test_path = f's3://{Bucket}/{Prefix}/dataset/test.csv'

train = pd.read_csv(train_path, index_col=0)
test  = pd.read_csv(test_path, index_col=0)

X_train = train.drop(columns='event')
y_train = train.get('event')

X_test  = test.drop(columns='event')
y_test  = test.get('event')

one_sample = X_test.sample(1).to_json(orient='records')
multiple_samples = X_test.sample(10).to_json(orient='records')

one_sample = ast.literal_eval(one_sample)       # format correctly to send
multiple_samples = json.loads(multiple_samples) # format correctly to send

#### Testing One Sample

import json

endpoint_name = 'ep-starbucks'

payload = json.dumps(one_sample)

response = sagemaker_runtime.invoke_endpoint(
    EndpointName=endpoint_name,
    ContentType='application/json',
    Body=payload
)

result = json.loads(response['Body'].read().decode())

print(f"Prediction of next events and probabilities: {result}")

#### Testing Multiple Samples

import json

endpoint_name = 'ep-starbucks'

payload = json.dumps(multiple_samples)

response = sagemaker_runtime.invoke_endpoint(
    EndpointName=endpoint_name,
    ContentType='application/json',
    Body=payload
)

result = json.loads(response['Body'].read().decode())

print(f"Prediction of next events and probabilities: {result}")

#### Predict over the enpoint all the test samples

endpoint_name = 'ep-starbucks'

batch_size=500
results = []
for i in range(0, len(X_test), batch_size):
    X_test_batch = X_test[i:i+batch_size].to_json(orient='records')
    X_test_batch = json.loads(X_test_batch)
    payload = json.dumps(X_test_batch)
    response = sagemaker_runtime.invoke_endpoint(EndpointName=endpoint_name, ContentType='application/json',Body=payload)
    result = json.loads(response['Body'].read().decode())
    result = pd.DataFrame(result)
    results.append(result)
    
results = pd.concat(results).reset_index(drop=True)
results

#event = results.columns

y_pred = results.apply(np.argmax, axis=1)#.apply(lambda x: event[x])

y = y_test.map(dict(zip(result.columns, range(4))))

pd.DataFrame(classification_report(y, y_pred, labels=range(4), target_names=results.columns, output_dict=True)).T

predictor.delete_endpoint()
